Fix Slow Loading on occurrence editor due to unperformant trait query (#2984)
# Summary
Fixes performance problem with getting traits per target collection.
Query was only not performing well when collid was being used in the
function due to having to use either collid as index.

Solution was to seperate the queries since I expect the total number of
unique state ids to remain relatively low meaning constructing a where
in will be a solid choice.

I should note that the query was slow regardess of whether their was
traits actually linked to occurrences or even without trait's being
enabled.