Adjust Material Sample language be be same as elsewhere (#3011)
* Adjust Material Sample language be be same as elsewhere

# Summary
`materialSampleVars.en.php` was not assigning to the $LANG array which
is not consitent with how we handle language loading. This difference
was breaking the lanugage loader and will break in Symbiota-Laravel
instances because it is introducing a previously unknown global variable

It also did not have translations for the fields which might need to
spun into a sperate issue.

* Add docs to new static fetch function

* Add es and fr translations

* Move language loading of ms labels to the function call so it is always loaded when called

---------

Co-authored-by: Katie Pearson <katelin.d.pearson24@gmail.com>